Correct variable names in calendar controller and add calendar styles
Fixed the incorrect parameter names in the calendar controller JavaScript file. There was a typo in 'locale' and the 'googleCalendar' parameters have been changed to use camel casing for consistency. Moreover, a new CSS class for 'calendar' has been added to the plugin.scss file to style the calendar interface. All these changes have also resulted in the update of the CSS and mix-manifest files.